Fixed buggy KBK mod file (now it works with NEURON v7.x as well).
The KBK mod file being used was the one originally included in the first
release of the Golgi cell model by Solinas et al on ModelDB, but later
modified because of incompatibility with NEURON v>6.2.

Note that the maximum conductance density for the channel on all cell
types had to be changed, too.